requirements.txt
streamlit==1.38.0
tensorflow==2.17.0
binance-connector==1.13.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.9.2
scikit-learn==1.5.1
plotly==5.22.0
tqdm==4.65.0
yfinance==0.2.61
